backend/requirements.txt
Gunicorn
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
Flask
Flask-Cors==3.0.10
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pycodestyle==2.10.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
